==> 02-intro.qmd <==
# Introduction {#sec-intro .unnumbered}

This pre-clinical analysis describes the analysis methods and data presentation to be used for the analysis and summarization of GLP/Non-GLP toxicity and biodistribution study in immunosuppressed and non-immunosuppressed cynomolgus monkeys, named as **"ToxBioDist (TBD) Analyzer"**.

The goal is to develop an integrative analysis workflow tailored to the unique challenges posed by data generated for genomic medicine, due to its multi-dimensional nature, i.e. across a number of platforms including vector genome copies (VGCs), mRNA, protein, and a wide range of responses. This workflow encompasses data structure standardization, quality checks, visualization of multi-dimensional data, identification of sources of variation, and quantification of exposure-response relationships, supporting posology assessments for pre-clinical to clinical program transitions. Specifically, this workflow is expected to be instrumental in analyzing biodistribution, expression of VGCs and mRNA, response, and toxicity, supporting the development of multiple programs.

With the recent acquisitions of genomic and cell therapy programs, Alexion CPSS designed and implemented a framework and an ops model for innovating toxicology and pathology expertise, building a state-of-the-art Quantitative Systems Modeling platform, providing immunogenicity sciences stewardship and leveraging the Data Sciences Unit for conducting risk: benefit and posology assessments for pre-clinical to clinical program transitions while collaborating with key stakeholders across Research and DRS organizations. This new framework expanded CPSS early-stage involvement, with minimal FTE increment, from CSID to TSID on programs and platform discovery projects to drive input in study designs, data structure standardization, quality checks, visualization of multi-dimensional unstructured data, identification of sources of variation, interpretation of NOAEL and safety margin estimations.

Alexion CPSS was able to meaningfully impact programs strategically at early stages and on into further stages including pre-IND. Proficiencies were developed with interpretation of in situ hybridization, immunohistochemistry, digital pathology, and evaluations of biodistribution and protein expression of vector genome copies (VGCs). Unique approaches to data visualization and interpretation were developed for biodistribution, expression of VGCs and mRNA, and toxicology study findings including histopathology. Our developed proficiencies and unique approaches have been demonstrated in moving our pipeline forward and addressing relevant agency questions effectively (e.g., ALXN2330, ALXN2340 and ALXN2350). The result was a strengthened focus on the 5Rs at the time of TSID (or earlier) continuing into later development.
