==> 05-biodist.qmd <==
# Biodistribution Data Analysis {#sec-biodist-analysis}

```{r}
#| results: "asis"
#| echo: false
addl.libpath = c("/opt/scp/services/azcore/biometrics/0.3.1/libraries/R/4.3.1" , 
                 "/opt/scp/services/azcore/corepackages/3.0.1/libraries/R/4.3.1",
                 "/opt/scp/services/azcore/coreutils/0.1.0/libraries/R/4.3.1" ,     
                 "/opt/scp/apps/gen-2021a/software/R/4.3.1-foss-2021a-core/lib64/R/site-library",
                 "/opt/scp/apps/gen-2021a/software/R/4.3.1-foss-2021a-core/lib64/R/library"
                 )

.libPaths(unique(c(  addl.libpath, .libPaths() )) )

source("_common.R")
load("./data/dat0_0118.RData")  # dat0, spec4matrix

library("kableExtra")
library("circlize")

```

Biodistribution and shedding studies are instrumental for developing adeno-associated virus (AAV)-based gene therapies. These studies play a crucial role in understanding the distribution and persistence of the viral vector within the body, as well as its potential dissemination into the environment.

The main objectives of biodistribution studies are to investigate where and in what form the vector is present within various tissues and biological fluids, and to determine whether measurable amounts of the vector are shed into the environment. Typically, polymerase chain reaction (PCR)-based bioanalytical methods are employed to measure AAV vector DNA in tissues and biological fluids. These methods provide sensitive and specific quantification of the vector, enabling researchers to assess the biodistribution profile and shedding patterns of the AAV vector.

In this analysis, we present the visualization of biodistribution and shedding data through various techniques, including scatter plots, line plots, bar plots, and heatmaps, to address the relevant questions.
