==> 05B-viral-shedding.qmd <==
```{r}
#| results: "asis"
#| echo: false

source("_common.R")
load("./data/dat0_0118.RData")  # dat0, spec4matrix

```

# Viral Shedding {#sec-viral-shedding-analysis}

## Background

Viral shedding refers to the release of virus particles from a host organism following administration. In the context of gene therapy, viral vectors are often used to deliver therapeutic genes to target cells. These vectors frequently include adeno-associated viruses (AAVs), lentiviruses, or retroviruses. While these vectors are designed to be safe and non-replicating, monitoring for viral shedding is essential for understanding their biodistribution and potential impact on non-target tissues or transmission to others.

**Expected Results**

1.  *Duration and Quantity*: Typically, viral shedding is expected to be transient, occurring shortly after administration and decreasing over time. The amount of shed virus is generally low, owing to the non-replicating nature of most viral vectors used in gene therapy.

2.  *Route of Shedding*: Shedding can occur through various biological fluids, such as blood, saliva, urine, or feces, depending on the vector type and administration route.

3.  *Biodistribution Patterns*: Minimal and transient presence in non-target tissues is anticipated due to the targeted nature of gene therapy vectors.

**Interpreting Viral Shedding Data**

-   *Safety Evaluation*: Low or undetectable levels of viral shedding indicate a favorable safety profile, reducing the risk of horizontal transmission or unintended effects on the environment.

-   *Vector Effectiveness*: Effectiveness is typically not directly linked to shedding levels. However, if shedding levels are unusually high, it might suggest incorrect targeting or immune clearance issues.

-   *Regulatory Compliance*: Monitoring and documentation of viral shedding are often required to fulfill regulatory guidelines and demonstrate the biosafety of gene therapy interventions.

-   *Duration Monitoring*: Monitoring should continue until shedding is no longer detectable, which helps establish the timeline for environmental containment and potential transmission to close contacts.

Viral shedding studies are crucial, not only for assessing the safety of the therapy for the individual patient but also for understanding potential environmental and public health impacts. If shedding is significant or prolonged, additional studies and measures may be necessary to mitigate risks.

## Observations

The viral shedding data is visualized in @fig-viral-shedding-facet-by-group-color-by-tissue. The data is faceted by treatment groups and colored by tissue type. The viral shedding data is also visualized in @fig-viral-shedding-facet-by-tissue-color-by-group, where the data is faceted by tissue type and colored by treatment groups. The viral shedding data in serum is visualized in @fig-viral-shedding-facet-by-tissue-color-by-group-serum, where the data is colored by treatment groups.

The data shows that viral shedding is minimal and transient, with a rapid decrease in concentration over time. The shedding is primarily observed in the blood and liver tissues, consistent with the expected biodistribution of the viral vector. The levels of shedding are low, indicating a low risk of transmission to other tissues or individuals. The data also demonstrates that the viral vector remains mostly contained within the target tissues, with minimal shedding in non-target tissues. These findings support the safety and efficacy of the gene therapy intervention.

<!-- ## Safety Profile -->

<!-- - **Low Risk of Transmission**: If shedding is minimal and transient, it suggests a low risk of transmission to others or to unintended tissues within the body. -->

<!-- - **Containment Assurance**: Limited shedding indicates that the viral vector remains mostly contained at the target site, minimizing environmental exposure and cross-contamination risks. -->

<!-- ## Efficacy Confirmation -->

<!-- - **Successful Targeting**: Expected shedding patterns confirm that the vector reaches and remains concentrated in the intended target tissues. -->

<!-- - **Reduced Off-target Effects**: Minimal shedding in non-target tissues supports the vector's specificity and reduces potential off-target effects. -->

<!-- ## Regulatory Compliance -->

<!-- - **Fulfillment of Guidelines**: Documented shedding within acceptable limits demonstrates adherence to regulatory safety guidelines and can aid in gaining approval from regulatory bodies. -->

<!-- - **Public Health Safety**: Data supports claims of minimal environmental and public health risk, ensuring safety beyond the direct recipient. -->

<!-- ## Potential for Further Development -->

<!-- - **Basis for Expanded Trials**: Positive shedding data might encourage further clinical trials or broader applications of the gene therapy. -->

<!-- - **Adjustments for Improvement**: If unexpected shedding patterns are observed, it may lead to adjustments in vector design or administration protocols to enhance safety and efficacy. -->

<!-- ## Need for Additional Monitoring -->

<!-- - If atypical or prolonged shedding is observed, further investigation may be recommended to understand and mitigate any unforeseen safety concerns. -->

<!-- These conclusions are integral to demonstrating both the safety and effectiveness of gene therapies, providing critical information for commercialization and broader clinical application. -->

```{r }
#| label: fun-viral-shedding
#| 
#' Plot Viral Shedding Data
#'
#' This function generates a plot for viral shedding data based on specified filters and parameters.
#'
#' @param data A data frame containing the viral shedding data. The data frame should include the following columns:
#'   - `ATPT` (numeric or character): Time point for the x-axis.
#'   - `ATPT_f` (factor or character): Time point factor to filter the data.
#'   - `AVAL` (numeric): The value to be plotted.
#'   - `GROUP_f` (factor or character): Group factor for faceting the plot.
#'   - `PARAMCD` (character): Parameter code to filter the data. 
#'   - `USUBJID` (character): Unique subject identifier.
#' @param PARAMCD_filter A character string specifying the parameter code to filter the data. Default is "PK".
#' @param ATPT_f_filter A vector specifying the time points to filter the data. Default is unique values of `ATPT_f` in the data.
#' @param y_scale_log10 A logical value indicating whether to use a log10 scale for the y-axis. Default is TRUE.
#'
#' @return A ggplot object representing the viral shedding data plot.
#'
#' @import dplyr
#' @import ggplot2
#' @import ggprism
#'
#' @examples
#' \dontrun{
#'   # Example usage:
#'   plot_virial_shedding_data(data = my_data, PARAMCD_filter = "PK", ATPT_f_filter = c("Day 1", "Day 2"), y_scale_log10 = TRUE)
#' }
#'
#' @export
 
 plot_viral_shedding_data <- function(data, 
                      PARAMCD_filter = "PK", 
                      ATPT_f_filter = unique(data$ATPT_f), 
                      
                      color_by = "TISSUE",
                      facet_by = "GROUP_f",
                      
                      y_scale_log10 = TRUE) {
  
  # The `df` variable is an internal working data frame used within each function.

  df <- data %>%  
    filter(PARAMCD == PARAMCD_filter) %>% 
    filter(ATPT_f %in% ATPT_f_filter) %>% 
    mutate(AVAL = as.numeric(AVAL)) %>% 
    drop_na(AVAL)  # Remove rows with NA in AVAL
  
    # Determine the range of the x-axis based on the filtered data
    x_limits <- range(df$ATPT, na.rm = TRUE)

  p <- ggplot(data = df, aes_string(x = "ATPT", y = "AVAL", color= color_by)) +
    geom_line(aes(group = USUBJID), color = rgb(0.5, 0.5, 0.5), size = 1, alpha = 0.3) + 
    geom_point() + 
    ggprism::theme_prism() +
    labs(
      x = "Time Point (Hrs)", 
      y = "Concentration (cp/µL)"#, 
      #title = paste("Plot for PARAMCD:", PARAMCD_filter)
    )    # Add labels and title
  
  if (y_scale_log10) {
    p <- p + scale_y_log10()  # Set y-axis to log10 scale
  }
  
  p <- p + 
    #facet_wrap(~GROUP_f) + 
    facet_wrap(as.formula(paste("~", facet_by))) #  ncol=1, scales="free_y")
  
    ggplot2::xlim(x_limits)  # Set x-axis limits based on the filtered data
  
  return(p)
}



#install.packages("testthat")
#library(testthat)

```

```{r}
#| label: fig-viral-shedding-facet-by-group-color-by-tissue
#| echo: false
#| fig-cap: |
#|   Kinetics of Viral Shedding Data Faceted by Treatment Groups and Colored by Tissue Type.
#| fig-cap-location: top
#| fig-width: 8
#| fig-height: 15  


#testthat::test_dir("tests/testthat")

plot_viral_shedding_data(
  dat0,  
  PARAMCD_filter = "PK"
) + 
  theme_prism2()  
  
  
 
  
  

 

```

```{r}
#| label: fig-viral-shedding-facet-by-tissue-color-by-group
#| echo: false
#| fig-cap: |
#|   Kinetics of Viral Shedding Data Faceted by Tissue Type and Colored by Treatment Groups.  
#| fig-cap-location: top
#| fig-width: 8
#| fig-height: 15  


plot_viral_shedding_data(
  dat0, #   
  PARAMCD_filter = "PK", 
  color_by = "GROUP_f ",
  facet_by = "TISSUE",
) + 
  theme_prism2() 

```

```{r}
#| label: fig-viral-shedding-facet-by-tissue-color-by-group-serum
#| echo: false
#| fig-cap: |
#|   Kinetics of Viral Shedding Data in Serum, Colored by Treatment Group.
#| fig-cap-location: top
#| fig-width: 8
#| fig-height: 15  


plot_viral_shedding_data(
    dat0 %>% filter(TISSUE == "Serum"), #   
    PARAMCD_filter = "PK", 
    color_by = "GROUP_f ",
    facet_by = "TISSUE",
) + 
    theme_prism2()  


```
